api_spot/requirements.txt
celery[redis]==5.3.4
Django==4.2.5
django-filter==23.2
django-phonenumber-field[phonenumberslite]==7.1.0
django-cors-headers==3.8.0
djangorestframework==3.14.0
djoser==2.2.0
drf-yasg==1.21.7
flower
gunicorn==21.2.0
phonenumbers==8.13.20
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0